Purchase service pattern implemented

+++ src/main/java/com/bookstoreapi/bookstoreapi/client/service/DeleteClientServiceImpl.java
@@ -2,7 +2,9 @@ package com.bookstoreapi.bookstoreapi.client.service;
 
 import com.bookstoreapi.bookstoreapi.client.Client;
 import com.bookstoreapi.bookstoreapi.client.ClientRepository;
+import com.bookstoreapi.bookstoreapi.purchase.service.PurchaseService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
@@ -12,10 +14,16 @@ public class DeleteClientServiceImpl implements DeleteClientService{
 
     @Autowired
     private ClientRepository clientRepository;
+    @Autowired
+    private PurchaseService purchaseService;
 
 
     @Override
     public void delete(Long id){
+        if(purchaseService.existsByBookId(id)){
+            throw new DataIntegrityViolationException(
+                    "One or more purchases have this client, it is not possible to delete");
+        }
         clientRepository.delete(this.findById(id));
     }
